Updated Metrics Structure
Moved all the previously defined modules into their own files to keep the code more seperated and cleaner
especially for future additions and changes which can now be made more local and dont all need to go into
the same ever growing file.
